pluggable architecture and refactoring
