+++ tests/conftest.py
@@ -1,4 +1,5 @@
 import pytest
+import mock
 
 
 pytest_plugins = 'pytester'
@@ -8,3 +9,14 @@ pytest_plugins = 'pytester'
 def variables(skin):
     return {'skins': {skin: {'base_url': 'http://',
                              'credentials': {}}}}
+
+
+@pytest.fixture
+def dummy_executor(parametrizer_class, navigation):
+    from pytest_play.engine import PlayEngine
+    return PlayEngine(navigation, {'foo': 'bar'}, parametrizer_class)
+
+
+@pytest.fixture
+def page_instance():
+    return mock.MagicMock()
