+++ tests/test_engine.py
@@ -4,21 +4,11 @@ from datetime import (
     datetime,
     timedelta,
 )
-from pytest_play.executors import JSONExecutorSplinter
 
 
-@pytest.fixture
-def page_instance():
-    return mock.MagicMock()
-
-
-@pytest.fixture
-def dummy_executor(parametrizer_class, navigation):
-    return JSONExecutorSplinter(navigation, {'foo': 'bar'}, parametrizer_class)
-
-
-def test_splinter_executor_constructor(bdd_vars, parametrizer_class):
-    executor = JSONExecutorSplinter(None, bdd_vars, parametrizer_class)
+def test_play_engine_constructor(bdd_vars, parametrizer_class):
+    from pytest_play.engine import PlayEngine
+    executor = PlayEngine(None, bdd_vars, parametrizer_class)
     assert executor.parametrizer_class is parametrizer_class
     assert executor.navigation is None
     assert executor.variables == bdd_vars
@@ -26,19 +16,6 @@ def test_splinter_executor_constructor(bdd_vars, parametrizer_class):
 
 def test_splinter_executor_parametrizer(dummy_executor):
     assert dummy_executor.parametrizer.parametrize('$foo') == 'bar'
-
-
-def test_splinter_executor_locator(dummy_executor):
-    assert dummy_executor.locator_translate(
-        {'type': 'css',
-         'value': 'body'}) == ('css', 'body')
-
-
-def test_splinter_executor_locator_bad(dummy_executor):
-    with pytest.raises(ValueError):
-        dummy_executor.locator_translate(
-            {'type': 'cssXX',
-             'value': 'body'}) == ('css', 'body')
 
 
 def test_splinter_execute(dummy_executor):
